added inference_pipeline
